requirements.txt
accelerate==0.29.3
argparse==1.4.0
ml-collections==0.1.1
nltk==3.8.1
scikit-image==0.21.0
scikit-learn==1.3.2
seaborn==0.12.2
timm==0.9.16
tqdm==4.66.3
transformers==4.40.0
wandb==0.15.0
datasets==3.6.0
huggingface_hub==0.24.0
evaluate
biopython==1.81
biotite==0.37.0
sentencepiece==0.1.99
hydra-core>=1.3
lightning
einops
numba
dm-tree
modelcif
omegaconf
h5py
lmdb
safetensors
py3Dmol
pyarrow
plotly
ninja
git+https://github.com/facebookresearch/esm.git@d7b3331f41442ed4ffde70cb95bdd48cabcec2e9
git+https://github.com/amyxlu/evo.git
git+https://github.com/MeshchaninovViacheslav/cheap-proteins.git
git+https://github.com/amyxlu/openfold.git